% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MOTE_data.R
\name{create.B}
\alias{create.B}
\title{Title
#TODO: improve the documentation}
\usage{
create.B(p, intercept = T)
}
\arguments{
\item{p}{The dimension of the design matrix}
}
\value{
A matrix
}
\description{
Title
#TODO: improve the documentation
}
\examples{
create.B(10)
}
