% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.MOTE}
\alias{importance.MOTE}
\alias{importance}
\title{MOTE variable importance}
\usage{
\method{importance}{MOTE}(x, ...)
}
\arguments{
\item{x}{MOTE object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of MOTE object.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=MOTE]{MOTE()}}
}
\author{
Marvin N. Wright
}
