% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTE.R
\name{MOTE}
\alias{MOTE}
\title{MOTE}
\usage{
MOTE(
  x.b,
  x.e,
  treat,
  y.b,
  y.e,
  Z,
  num.trees = 300,
  min.node.size = NULL,
  max.depth = NULL,
  num.random.splits = 1,
  minprop = 0.1,
  write.forest = TRUE,
  replace = TRUE,
  sample.fraction = ifelse(replace, 1, 0.632),
  class.weights = NULL,
  keep.inbag = FALSE,
  inbag = NULL,
  holdout = FALSE,
  oob.error = TRUE,
  num.threads = 1,
  verbose = TRUE,
  seed = NULL,
  dependent.variable.name = NULL,
  trt.variable.name = NULL,
  ...
)
}
\arguments{
\item{num.trees}{Number of trees.}

\item{min.node.size}{Minimal node size. Default 1 for classification, 5 for regression, 3 for survival, and 10 for probability.}

\item{max.depth}{Maximal tree depth. A value of NULL or 0 (the default) corresponds to unlimited depth, 1 to tree stumps (1 split per tree).}

\item{num.random.splits}{For "extratrees" splitrule.: Number of random splits to consider for each candidate splitting variable.}

\item{minprop}{For "maxstat" splitrule: Lower quantile of covariate distribution to be considered for splitting.}

\item{write.forest}{Save \code{MOTE.forest} object, required for prediction. Set to \code{FALSE} to reduce memory usage if no prediction intended.}

\item{replace}{Sample with replacement.}

\item{sample.fraction}{Fraction of observations to sample. Default is 1 for sampling with replacement and 0.632 for sampling without replacement. For classification, this can be a vector of class-specific values.}

\item{class.weights}{Weights for the outcome classes (in order of the factor levels) in the splitting rule (cost sensitive learning). Classification and probability prediction only. For classification the weights are also applied in the majority vote in terminal nodes.}

\item{keep.inbag}{Save how often observations are in-bag in each tree.}

\item{inbag}{Manually set observations per tree. List of size num.trees, containing inbag counts for each observation. Can be used for stratified sampling.}

\item{holdout}{Hold-out mode. Hold-out all samples with case weight 0 and use these for variable importance and prediction error.}

\item{oob.error}{Compute OOB prediction error. Set to \code{FALSE} to save computation time, e.g. for large survival forests.}

\item{num.threads}{Number of threads. Default is number of CPUs available.}

\item{verbose}{Show computation status and estimated runtime.}

\item{seed}{Random seed. Default is \code{NULL}, which generates the seed from \code{R}. Set to \code{0} to ignore the \code{R} seed.}

\item{dependent.variable.name}{Name of dependent variable, needed if no formula given. For survival forests this is the time variable.}

\item{trt.variable.name}{Name of treatment variable. Treatment variable should be a factor variable}

\item{...}{Further arguments passed to or from other methods (currently ignored).}

\item{formula}{Object of class \code{formula} or \code{character} describing the model to fit. Interaction terms supported only for numerical variables.}

\item{data}{Training data of class \code{data.frame}, \code{matrix}, \code{dgCMatrix} (Matrix).}

\item{save.memory}{Use memory saving (but slower) splitting mode. No effect for survival and GWAS data. Warning: This option slows down the tree growing, use only if you encounter memory problems.}

\item{x}{Predictor data (independent variables), alternative interface to data with formula or dependent.variable.name.}

\item{y}{Response vector (dependent variable), alternative interface to data with formula or dependent.variable.name. For survival use a \code{Surv()} object or a matrix with time and status.}
}
\value{
Object of class \code{ranger} with elements
\item{\code{forest}}{Saved forest (If write.forest set to TRUE). Note that the variable IDs in the \code{split.varIDs} object do not necessarily represent the column number in R.}
\item{\code{predictions}}{Predicted classes/values, based on out of bag samples (classification and regression only).}
\item{\code{variable.importance}}{Variable importance for each independent variable.}
\item{\code{prediction.error}}{Overall out of bag prediction error. For classification this is the fraction of missclassified samples, for probability estimation the Brier score, for regression the mean squared error and for survival one minus Harrell's C-index.}
\item{\code{r.squared}}{R squared. Also called explained variance or coefficient of determination (regression only). Computed on out of bag data.}
\item{\code{confusion.matrix}}{Contingency table for classes and predictions based on out of bag samples (classification only).}
\item{\code{call}}{Function call.}
\item{\code{num.trees}}{Number of trees.}
\item{\code{num.independent.variables}}{Number of independent variables.}
\item{\code{mtry}}{Value of mtry used.}
\item{\code{min.node.size}}{Value of minimal node size used.}
\item{\code{num.samples}}{Number of samples.}
\item{\code{inbag.counts}}{Number of times the observations are in-bag in the trees.}
}
\description{
MOTE is a fast implementation of MOTE RF (Guo et al. 2021), a random forest model that simultaneously estimates individualized treatment effects of multivariate outcomes.
}
\details{
Forests of extremely randomized trees (Geurts et al. 2006) can be grown.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Note that:
\itemize{
\item nodes with size smaller than \code{min.node.size} can occur, as in original Random Forests.
\item The mtry variable that randomly selected variables doesn't exist in our framework, as we are using oblique splitting rules.
\item For factor covriates, a reference coding design matrix will be created accordingly. TODO: it is possible to extend this later by extending to ordered factors.
\item The importance of variables/features are calculated by accumulating the coefficients in the oblique splitting rules weighted by node sample size.
\item The model yet to provide handling of missing value. Data containing missing value can promote computation error and warning.
}

For a large number of variables and data frames as input data the formula interface can be slow or impossible to use.
TODO: need improve this part a little bit
Alternatively \code{dependent.variable.name} and \code{status.variable.name} for treatment category or \code{x} and \code{y} can be used.
Use \code{x} and \code{y} with a matrix for \code{x} to avoid conversions and save memory.

TODO:
Since our algorithm is oriented to analyze microbiome data, we can directly use phyloseq object.
Case weights is not supported temporarily

See \url{https://github.com/boyiguo1/MOTE.RF} for the development version.

With recent R versions, multithreading on Windows platforms should just work.
If you compile yourself, the new RTools toolchain is required.
}
\examples{
set.seed(1)

p = 10
q = 3

B = create.B(p)
Z = create.Z(p,q)

sim.dat <- sim_MOTE_data(B = B, Z = Z)





}
\references{
\itemize{
\item Guo et al. (Submitted). Estimating Heterogeneous Treatment Effect on Multivariate Responses using Random Forests. \url{https://doi.org/xxxx}.
}
}
\seealso{
\code{\link[=predict.MOTE]{predict.MOTE()}}
}
\author{
Boyi Guo
}
