% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOTE.RF.R
\docType{package}
\name{MOTE.RF}
\alias{MOTE.RF}
\title{MOTE.RF: A package for estimating heterogeneous treatment effect on multivariate responses using random forests}
\description{
// TODO: Update this page
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

